+++ molviewspec/molviewspec/nodes.py
@@ -1392,6 +1392,7 @@ class CameraParams(BaseModel):
     up: Vec3[float] = Field(
         description="Controls the rotation around the vector between target and position", required=True
     )
+    near: Optional[float] = Field(None, description="Near clipping plane distance from the position")
 
 
 class CanvasParams(BaseModel):