+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file, following t
 ## [Unreleased]
 
 - Add support for `dx` volume format
+- Add support for clipping of primitives
+- Add support for `near` in `CameraParams`
 
 ## [v1.7.0] - 2025-04-22